DTO 별점 자료형 Float -> Integer 변환

+++ src/main/java/com/arton/backend/performance/applicaiton/data/PerformanceDetailQueryDslDtoV3.java
@@ -137,7 +137,7 @@ public class PerformanceDetailQueryDslDtoV3 {
                 .ticketOpenDate(StringUtils.hasText(ticketOpenDay) ? ticketOpenDay : "미정")
                 .ticketEndDate(StringUtils.hasText(ticketEndDay) ? ticketEndDay : "미정")
                 .isZzim(zzim == null ? false : zzim)
-                .startScore(getStartScore() == null ? 0f : getStartScore())
+                .startScore(getStartScore() == null ? 0 : getStartScore().intValue())
                 .reviewCount(getReviewCount() == null ? 0 : getReviewCount())
                 .build();
     }
